requirements.txt
boto3~=1.36.6
python-dotenv~=1.0.1
fastapi~=0.115.7